requirements.txt
Django==1.6.1
Fabric==1.8.1
MySQL-python==1.2.5
Pillow==2.3.0
South==0.8.4
Unidecode==0.04.14
amqp==1.3.3
anyjson==0.3.3
argparse==1.2.1
beautifulsoup4==4.3.2
billiard==3.3.0.13
biplist==0.6
bitly-api==0.3
celery==3.1.7
celery-haystack==0.7.2
chardet==2.2.1
chromium-compact-language-detector==0.031415
defusedxml==0.4.1
django-appconf==0.6
django-bootstrap-toolkit==2.15.0
django-celery==3.1.1
django-celery-transactions==0.1.3
django-celery-with-redis==3.0
django-crispy-forms==1.4.0
django-debug-toolbar==1.0.1
django-facebook==5.3.1
django-haystack==2.1.0
django-oauth2-provider==0.2.6
django-taggit==0.11.2
django-tastypie==0.11.0
dnspython==1.11.1
ecdsa==0.10
eventlet==0.14.0
evernote==1.25.0
feedparser==5.1.3
gevent==1.0
greenlet==0.4.2
grequests==0.2.0
gunicorn==18.0
hiredis==0.1.2
httplib2==0.8
itsdangerous==0.23
jsonfield==0.9.20
kombu==3.0.8
lxml==3.2.4
mimeparse==0.1.3
nydus==0.10.6
oauth2==1.5.211
oauthlib==0.6.0
paramiko==1.12.1
pycrypto==2.6.1
pyelasticsearch==0.6.1
pylibmc==1.2.3
python-dateutil==2.2
python-mimeparse==0.1.4
python-openid==2.2.5
pytz==2013.9
redis==2.9.0
requests==2.1.0
shortuuid==0.4
simplejson==3.3.2
six==1.5.2
sphinx-me==0.2.1
sqlparse==0.1.10
wsgiref==0.1.2
